[Refactor] change file name to follow naming conventions
