+++ backend/auth/authentication.js
@@ -4,7 +4,7 @@ const bcrypt = require('bcrypt')
 const jwt = require('jsonwebtoken')
 const config = require('config')
 
-const User = require('../models/user')
+const User = require('../models/user.model')
 
 const generateAuthToken = user => {
   const token = jwt.sign({ _id: user._id }, config.get('JWT_SECRET'), {
